data-processing/requirements.txt
pandas==2.1.2
numpy==1.26.1
sanic==23.6.0
sanic_cors==2.2.0
aiohttp==3.8.6
ulid==1.1
minio==7.1.17
zhipuai==1.0.7
langchain==0.0.336
spacy==3.5.4
pypdf==3.17.1
emoji==2.2.0
ftfy==6.1.1
psycopg2_binary==2.9.9
kubernetes==25.3.0
duckdb==0.9.2
DBUtils==3.0.3
pyyaml==6.0.1

opencc==0.2
opencc-python-reimplemented==0.1.7
selectolax==0.3.17
